Port interrupt fix to release branch (#8961)
* Fix interrupting on Windows (#8949)

* Try adding win32-api (not sure it will build on Unix)

* Working interrupt on windows

* Use a daemon to do the interrupting so we don't have to wait for a process to start

* Add windows non conda test

* Add news entry

* Update 3rd party notices

* Fix unit test

* Remove logging to my own machine

* Fix build and update changelog

* Remove news file

* Put changelog back

+++ src/client/datascience/kernel-launcher/kernelProcess.ts
@@ -6,6 +6,7 @@ import { ChildProcess } from 'child_process';
 import { kill } from 'process';
 import * as fs from 'fs-extra';
 import * as tmp from 'tmp';
+import * as os from 'os';
 import { CancellationToken, Event, EventEmitter } from 'vscode';
 import { IPythonExtensionChecker } from '../../api/types';
 import { CancellationError, createPromiseFromCancellation } from '../../common/cancellation';
@@ -26,7 +27,7 @@ import { createDeferred } from '../../common/utils/async';
 import * as localize from '../../common/utils/localize';
 import { noop, swallowExceptions } from '../../common/utils/misc';
 import { captureTelemetry } from '../../telemetry';
-import { Telemetry } from '../constants';
+import { KernelInterruptDaemonModule, Telemetry } from '../constants';
 import {
     connectionFilePlaceholder,
     findIndexOfConnectionFile,
@@ -42,6 +43,7 @@ import { PythonKernelDiedError } from '../errors/pythonKernelDiedError';
 import { KernelDiedError } from '../errors/kernelDiedError';
 import { KernelPortNotUsedTimeoutError } from '../errors/kernelPortNotUsedTimeoutError';
 import { ignoreLogging, TraceOptions } from '../../logging/trace';
+import { PythonKernelInterruptDaemon } from './pythonKernelInterruptDaemon';
 
 // Launches and disposes a kernel process given a kernelspec and a resource or python interpreter.
 // Exposes connection information and the process itself.
@@ -67,11 +69,13 @@ export class KernelProcess implements IKernelProcess {
         return true;
     }
     private _process?: ChildProcess;
+    private _interruptDaemon?: PythonKernelInterruptDaemon;
     private exitEvent = new EventEmitter<{ exitCode?: number; reason?: string }>();
     private launchedOnce?: boolean;
     private disposed?: boolean;
     private connectionFile?: string;
     private _launchKernelSpec?: IJupyterKernelSpec;
+    private _interruptSignalHandle = 0;
     private readonly _kernelConnectionMetadata: Readonly<
         LocalKernelSpecConnectionMetadata | PythonKernelConnectionMetadata
     >;
@@ -90,9 +94,20 @@ export class KernelProcess implements IKernelProcess {
     public async interrupt(): Promise<void> {
         if (!this.canInterrupt) {
             throw new Error('Kernel interrupt not supported in KernelProcess.ts');
-        } else if (this._kernelConnectionMetadata.kernelSpec.interrupt_mode !== 'message' && this._process) {
-            traceInfo('Interrupting kernel via Signals');
+        } else if (
+            this._kernelConnectionMetadata.kernelSpec.interrupt_mode !== 'message' &&
+            this._process &&
+            !this._interruptSignalHandle
+        ) {
+            traceInfo('Interrupting kernel via SIGINT');
             kill(this._process.pid, 'SIGINT');
+        } else if (
+            this._kernelConnectionMetadata.kernelSpec.interrupt_mode !== 'message' &&
+            this._process &&
+            this._interruptSignalHandle
+        ) {
+            traceInfo('Interrupting kernel via custom event (Win32)');
+            return this.fireWin32InterruptEvent();
         } else {
             traceError('No process to interrupt in KernleProcess.ts');
         }
@@ -257,6 +272,7 @@ export class KernelProcess implements IKernelProcess {
         traceInfo('Dispose Kernel process');
         this.disposed = true;
         swallowExceptions(() => {
+            void this._interruptDaemon?.kill();
             this._process?.kill(); // NOSONAR
             this.exitEvent.fire({});
         });
@@ -389,7 +405,7 @@ export class KernelProcess implements IKernelProcess {
                 interpreter: this._kernelConnectionMetadata.interpreter
             });
 
-            const [executionService, wdExists, env] = await Promise.all([
+            let [executionService, wdExists, env] = await Promise.all([
                 executionServicePromise,
                 fs.pathExists(workingDirectory),
                 this.kernelEnvVarsService.getEnvironmentVariables(
@@ -398,6 +414,15 @@ export class KernelProcess implements IKernelProcess {
                     this._kernelConnectionMetadata.kernelSpec
                 )
             ]);
+
+            // On windows, in order to support interrupt, we have to set an environment variable pointing to a WIN32 event handle
+            if (os.platform() === 'win32') {
+                env = env || process.env;
+                const handle = await this.getWin32InterruptHandle();
+
+                // See the code ProcessPollingWindows inside of ipykernel for it listening to this event handle.
+                env.JPY_INTERRUPT_EVENT = `${handle}`;
+            }
 
             // If we don't have a KernelDaemon here & we're not running a Python module either.
             // The kernelspec argv could be something like [python, main.py, --something, --something-else, -f,{connection_file}]
@@ -456,5 +481,30 @@ export class KernelProcess implements IKernelProcess {
 
         this._process = exeObs.proc;
         return exeObs;
+    }
+
+    private async getWin32InterruptHandle(): Promise<number> {
+        if (!this._interruptSignalHandle) {
+            const interruptDaemon = await this.pythonExecFactory.createDaemon({
+                daemonModule: KernelInterruptDaemonModule,
+                resource: this.resource,
+                interpreter: this._kernelConnectionMetadata.interpreter!,
+                daemonClass: PythonKernelInterruptDaemon,
+                dedicated: true
+            });
+            if ('kill' in interruptDaemon) {
+                this._interruptDaemon = interruptDaemon;
+            }
+            if (this._interruptDaemon) {
+                this._interruptSignalHandle = await this._interruptDaemon.getInterruptHandle();
+            }
+        }
+        return this._interruptSignalHandle;
+    }
+
+    private async fireWin32InterruptEvent() {
+        if (this._interruptDaemon) {
+            return this._interruptDaemon.interrupt();
+        }
     }
 }
